tired

+++ torch_points3d/core/data_transform/transforms.py
@@ -521,7 +521,7 @@ class ShiftVoxels:
                 raise Exception("should quantize first using GridSampling3D")
             if not isinstance(data.coords, torch.IntTensor):
                 raise Exception("The pos are expected to be coordinates, so torch.IntTensor")
-            data.coords[:, :3] += (torch.rand(3) * 100).type_as(data.pos)
+            data.coords[:, :3] += (torch.rand(3) * 100).type_as(data.coords)
         return data
 
     def __repr__(self):
